zeenode/requirements.txt
discord.py
colorama
aiohttp
requests
pyfiglet
datetime
urllib3
asyncio
